Relocate all basic post process passes
