+++ multipacks-engine/src/main/java/multipacks/postprocess/basic/IncludePass.java
@@ -13,35 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package multipacks.postprocess;
+package multipacks.postprocess.basic;
 
 import java.io.IOException;
 
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
 import multipacks.bundling.BundleResult;
+import multipacks.postprocess.PostProcessPass;
 import multipacks.utils.Selects;
 import multipacks.utils.logging.AbstractMPLogger;
 import multipacks.vfs.Path;
 import multipacks.vfs.VirtualFs;
 
-public class MovePass extends PostProcessPass {
-	private Path from;
-	private Path to;
+public class IncludePass extends PostProcessPass {
+	private Path file;
 
-	public MovePass(JsonObject config) {
-		from = new Path(Selects.nonNull(config.get("from"), "'from' is empty").getAsString());
-		to = new Path(Selects.nonNull(config.get("to"), "'to' is empty").getAsString());
+	public IncludePass(JsonObject config) {
+		file = new Path(Selects.nonNull(config.get("file"), "'file' is empty").getAsString());
 	}
 
 	@Override
 	public void apply(VirtualFs fs, BundleResult result, AbstractMPLogger logger) throws IOException {
-		for (Path pFrom : fs.ls(from)) {
-			String tail = pFrom.toString().substring(from.toString().length() + 1);
-			Path pTo = Path.join(to, new Path(tail));
-			byte[] bs = fs.read(pFrom);
-			fs.write(pTo, bs);
-			fs.delete(pFrom);
-		}
+		JsonElement json = fs.readJson(file);
+		PostProcessPass.apply(json, fs, result, logger);
 	}
 }
